⚒️ Forge: [ferrous-chimera refactor] (#4338)
🚮 Smell: `experiments/ferrous-chimera` had "God functions" (like `act` with repetitive stack popping), duplicated bounds-checking/wrap-around logic, redundant casts, and unused structural fields (`energy`, `mass`, `radius`, `with_velocity`) that generated dead code warnings.
✨ Solution: Extracted coordinate mapping into `Universe::world_to_grid`, removed dead code structurally rather than hiding it with `#[allow(dead_code)]`, created `Body::apply_wrap_around` to deduplicate wrap logic, and pulled stack popping into `Agent::pop_clamped`. Replaced modulo math with `is_multiple_of(5)` and stripped out redundant `as f64` casts.
🧼 Benefit: Reduced code noise, flattened control flow, eradicated warnings correctly, and improved maintainability.
🛡️ Verification: Tests passed. `cargo clippy` is green. No logic changed.

+++ experiments/ferrous-chimera/src/main.rs
@@ -155,14 +155,8 @@ where
                     "Ferrous Chimera | Agents: {} | Magnetic Stigmergy Active",
                     universe.agents.len()
                 )))
-                .x_bounds([
-                    pan.x as f64 - 100.0 * zoom as f64,
-                    pan.x as f64 + 100.0 * zoom as f64,
-                ])
-                .y_bounds([
-                    pan.y as f64 - 100.0 * zoom as f64,
-                    pan.y as f64 + 100.0 * zoom as f64,
-                ])
+                .x_bounds([pan.x - 100.0 * zoom, pan.x + 100.0 * zoom])
+                .y_bounds([pan.y - 100.0 * zoom, pan.y + 100.0 * zoom])
                 .paint(|ctx| {
                     // Draw Platter (Magnetism)
                     let step = 4; // Faster render
@@ -189,10 +183,10 @@ where
                         // Trail
                         for i in 0..agent.body.trail.len().saturating_sub(1) {
                             ctx.draw(&Line {
-                                x1: agent.body.trail[i].x as f64,
-                                y1: agent.body.trail[i].y as f64,
-                                x2: agent.body.trail[i + 1].x as f64,
-                                y2: agent.body.trail[i + 1].y as f64,
+                                x1: agent.body.trail[i].x,
+                                y1: agent.body.trail[i].y,
+                                x2: agent.body.trail[i + 1].x,
+                                y2: agent.body.trail[i + 1].y,
                                 color: Color::Gray,
                             });
                         }
@@ -200,8 +194,8 @@ where
                         // Body
                         let symbol = if agent.vm.halted { "X" } else { "O" };
                         ctx.print(
-                            agent.body.pos.x as f64,
-                            agent.body.pos.y as f64,
+                            agent.body.pos.x,
+                            agent.body.pos.y,
                             Span::styled(symbol, Style::default().fg(agent.body.color)),
                         );
                     }
